refactor : css 스타일 분리
