+++ src/component/Body/Body.tsx
@@ -1,14 +1,4 @@
-import { css } from "@emotion/react";
-
-const bodyLayout = css`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: 36px 25px;
-  box-sizing: border-box;
-  gap: 28px;
-  min-height: 100vh;
-`;
+import { bodyLayout } from "./Body.style";
 
 export default function Body({ children }: { children: React.ReactNode }) {
   return <body css={bodyLayout}>{children}</body>;
