+++ src/component/Button/Button.tsx
@@ -1,4 +1,4 @@
-import { css } from "@emotion/react";
+import { buttonLayout } from "./Button.style";
 
 type ButtonStyleProps = "primary" | "secondary";
 
@@ -7,21 +7,6 @@ interface ButtonProps {
   children: React.ReactNode;
   style?: ButtonStyleProps;
 }
-
-const buttonLayout = (backgroundColor: string, color: string) => {
-  return css`
-    display: flex;
-    align-items: center;
-    padding: 4px 8px;
-    background-color: ${backgroundColor};
-    border-radius: 4px;
-    color: ${color};
-    gap: 4px;
-    font-weight: 600;
-    font-family: "Noto Sans";
-    font-size: 12px;
-  `;
-};
 
 export default function Button({
   onClick,
