+++ src/component/Header/Header.style.ts
@@ -1,0 +1,19 @@
+import { css } from "@emotion/react";
+
+const headerLayout = css`
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 24px;
+  box-sizing: border-box;
+  width: 100%;
+  height: 64px;
+  background-color: black;
+  color: #fff;
+  font-family: "Noto Sans";
+  font-size: 20px;
+  font-weight: 800;
+`;
+
+export { headerLayout };
