+++ src/component/Header/Header.tsx
@@ -1,25 +1,9 @@
-import { css } from "@emotion/react";
+import { headerLayout } from "./Header.style";
 
 interface HeaderProps {
   title: string;
   children?: React.ReactNode;
 }
-
-const headerLayout = css`
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 24px;
-  box-sizing: border-box;
-  width: 100%;
-  height: 64px;
-  background-color: black;
-  color: #fff;
-  font-family: "Noto Sans";
-  font-size: 20px;
-  font-weight: 800;
-`;
 
 export default function Header({ title, children }: HeaderProps) {
   return (
