+++ src/component/Product/Product.tsx
@@ -1,7 +1,14 @@
-import { css } from "@emotion/react";
 import { deleteCartItem, postCartItem } from "../../api/cartItem";
 import { CartItem } from "../../page/ShopPage";
 import Button from "../Button/Button";
+import {
+  contentLayout,
+  descriptionLayout,
+  imgLayout,
+  priceLayout,
+  productLayout,
+  productNameLayout,
+} from "./Product.style";
 
 interface ProductProps {
   id: string;
@@ -11,51 +18,6 @@ interface ProductProps {
   selectedCardItems: CartItem[];
   onChange: () => void;
 }
-
-const productLayout = css`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-bottom: 8px;
-  border-radius: 8px;
-  width: 182px;
-  height: 224px;
-  gap: 15px;
-`;
-
-const imgLayout = css`
-  border-radius: 8px 8px 0 0;
-  width: 100%;
-  height: 50%;
-`;
-
-const contentLayout = css`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-end;
-  gap: 27px;
-  width: 100%;
-  height: 50%;
-`;
-
-const descriptionLayout = css`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 8px;
-  width: 100%;
-`;
-
-const productNameLayout = css`
-  font-size: 14px;
-  font-weight: 700;
-`;
-
-const priceLayout = css`
-  font-size: 12px;
-  font-weight: 500;
-`;
 
 export default function Product({
   id,
