+++ src/component/ProductContainer/ProductContainer.style.ts
@@ -1,0 +1,9 @@
+import { css } from "@emotion/react";
+
+const ProductContainerLayout = css`
+  display: grid;
+  justify-content: space-between;
+  grid-template-columns: repeat(2, auto);
+`;
+
+export { ProductContainerLayout };
