+++ src/component/ProductContainer/ProductContainer.tsx
@@ -1,6 +1,6 @@
-import { css } from "@emotion/react";
 import { CartItem } from "../../page/ShopPage";
 import Product from "../Product/Product";
+import { ProductContainerLayout } from "./ProductContainer.style";
 
 export interface Product {
   id: string;
@@ -14,12 +14,6 @@ interface ProductContainerProps {
   cartItemList: CartItem[];
   onChange: () => void;
 }
-
-const ProductContainerLayout = css`
-  display: grid;
-  justify-content: space-between;
-  grid-template-columns: repeat(2, auto);
-`;
 
 export default function ProductContainer({
   products,
