+++ src/component/TitleContainer/TitleContainer.style.ts
@@ -1,0 +1,14 @@
+import { css } from "@emotion/react";
+
+const titleContainerLayout = css`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  background-color: white;
+  font-weight: 700;
+  font-size: 24px;
+  line-height: 100%;
+  gap: 24px;
+`;
+
+export { titleContainerLayout };
