+++ src/component/Toast/Toast.style.ts
@@ -1,0 +1,17 @@
+import { css } from "@emotion/react";
+
+const ToastLayout = css`
+  position: absolute;
+  bottom: -36px;
+  left: 0;
+  padding: 12px 77px;
+  width: 100%;
+  background-color: #ffc9c9;
+  font-weight: 500;
+  font-size: 12px;
+  text-align: center;
+  color: black;
+  box-sizing: border-box;
+`;
+
+export { ToastLayout };
